requirements.txt
datasets==4.4.1
gradio==6.2.0
lightning==2.5.6
numpy==2.4.0
Pillow==12.1.0
torch==2.9.0
torchmetrics==1.7.4
torchvision==0.24.0
